PR Fix, moving schemas.js -> lib/collections
