+++ client/index.js
@@ -2,4 +2,4 @@ import "./templates";
 import "./defaults";
 import "./init";
 import "./styles/layout.less";
-import "../lib/schemas.js"
+import "../lib/collections/schemas.js"
